automated/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
Pillow==10.2.0
ExifRead==3.0.0
python-magic==0.4.27
django-cors-headers==4.3.1
PyExifTool==0.5.6
python-decouple==3.8
